reorder moving users between rooms
moving a user from the waiting room to a task room and then removing that
user from the waiting room causes a disconnect in the task room.

this commit first removes the user from the waiting room and only
afterwards moves her to the task room. no disconnect happens in that case.